move app setting context items from bootstrapping to own context
