+++ src/renderer/contexts/main/AppSettings/defaults.ts
@@ -1,0 +1,15 @@
+// Copyright 2024 @rossbulat/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
+
+import type { AppSettingsContextInterface } from './types';
+
+export const defaultAppSettingsContext: AppSettingsContextInterface = {
+  dockToggled: true,
+  silenceOsNotifications: false,
+  showDebuggingSubscriptions: false,
+  setSilenceOsNotifications: (b) => {},
+  handleDockedToggle: () => {},
+  handleToggleSilenceOsNotifications: () => {},
+  handleToggleShowDebuggingSubscriptions: () => {},
+};
