+++ src/renderer/contexts/main/Bootstrapping/default.ts
@@ -9,13 +9,6 @@ export const defaultBootstrappingContext: BootstrappingInterface = {
   online: false,
   isAborting: false,
   isConnecting: false,
-  dockToggled: true,
-  silenceOsNotifications: false,
-  showDebuggingSubscriptions: false,
-  setSilenceOsNotifications: (b) => {},
-  handleToggleSilenceOsNotifications: () => {},
-  handleToggleShowDebuggingSubscriptions: () => {},
-  handleDockedToggle: () => {},
   setAppLoading: (b) => {},
   setIsAborting: (b) => {},
   setIsConnecting: (b) => {},
