+++ src/renderer/contexts/main/Bootstrapping/types.ts
@@ -8,13 +8,6 @@ export interface BootstrappingInterface {
   isAborting: boolean;
   isConnecting: boolean;
   online: boolean;
-  dockToggled: boolean;
-  silenceOsNotifications: boolean;
-  showDebuggingSubscriptions: boolean;
-  setSilenceOsNotifications: (b: boolean) => void;
-  handleDockedToggle: () => void;
-  handleToggleSilenceOsNotifications: () => void;
-  handleToggleShowDebuggingSubscriptions: () => void;
   setAppLoading: (b: boolean) => void;
   setIsAborting: (b: boolean) => void;
   setIsConnecting: (b: boolean) => void;
