+++ src/renderer/library/Header/index.tsx
@@ -1,19 +1,20 @@
 // Copyright 2024 @rossbulat/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import { ButtonSecondary } from '@app/kits/Buttons/ButtonSecondary';
+import { Config as RendererConfig } from '@/config/processes/renderer';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
+import { faUnlock, faLock } from '@fortawesome/pro-solid-svg-icons';
+import { HeaderWrapper } from './Wrapper';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Menu } from '@app/library/Menu';
+import { useAppSettings } from '@/renderer/contexts/main/AppSettings';
 import { useLocation } from 'react-router-dom';
-import { HeaderWrapper } from './Wrapper';
-import { ButtonSecondary } from '@app/kits/Buttons/ButtonSecondary';
-import { useBootstrapping } from '@app/contexts/main/Bootstrapping';
-import { Config as RendererConfig } from '@/config/processes/renderer';
-import { faUnlock, faLock } from '@fortawesome/pro-solid-svg-icons';
 import type { HeaderProps } from './types';
 
 export const Header = ({ showMenu, appLoading = false }: HeaderProps) => {
   const { pathname } = useLocation();
+  const { dockToggled, handleDockedToggle } = useAppSettings();
 
   /// Determine active window by pathname.
   let activeWindow: string;
@@ -25,13 +26,10 @@ export const Header = ({ showMenu, appLoading = false }: HeaderProps) => {
       activeWindow = 'menu';
   }
 
-  const { dockToggled, handleDockedToggle } = useBootstrapping();
-
   /// Handle clicking the docked button.
   const handleDocked = () => {
     handleDockedToggle();
 
-    // Post message to settings window to update switch.
     RendererConfig.portToSettings.postMessage({
       task: 'settings:set:dockedWindow',
       data: {
